Remove `|| 'All'`s from template.
Two of them were using incorrect OR, @foo.join(',') | 'All'

The single bar doesn't work with Strings in ruby. However, since
non-null defaults are set in the sudoers class, it doesn't seem like
that second case will fire unless someone very explicitly sets them to
null. And if they do, so be it.